tests for controllers added

+++ src/main/java/com/example/user_registration/web/CompanyController.java
@@ -2,13 +2,15 @@ package com.example.user_registration.web;
 
 import com.example.user_registration.model.Company;
 import com.example.user_registration.repo.CompanyRepo;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequestMapping("/companies")
-public class CompanyController { // (!) still available by usual User
+public class CompanyController {
     private CompanyRepo companyRepo;
 
     public CompanyController(CompanyRepo companyRepo) {
@@ -17,12 +19,18 @@ public class CompanyController { // (!) still available by usual User
 
     @GetMapping
     public String listCompanies(Model model) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         model.addAttribute("companiesList", companyRepo.findAll());
         return "companies";
     }
 
     @GetMapping("/add")
     public String addCompany(Model model) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         model.addAttribute("company", new Company());
         model.addAttribute("errMsg", "");
         return "create_company";
@@ -30,6 +38,9 @@ public class CompanyController { // (!) still available by usual User
 
     @PostMapping("/add")
     public String createCompany(@ModelAttribute("company") Company company, Model model) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         if (companyRepo.findByName(company.getName()) != null) {
             model.addAttribute("errMsg", "Error: company with such name already exists");
             return "create_company";
@@ -41,6 +52,9 @@ public class CompanyController { // (!) still available by usual User
 
     @GetMapping("/edit/{id}")
     public String editCompany(@PathVariable("id") Long id, Model model) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         model.addAttribute("company", companyRepo.findById(id));
         model.addAttribute("errMsg", "");
         return "edit_company";
@@ -48,6 +62,9 @@ public class CompanyController { // (!) still available by usual User
 
     @PostMapping("/edit/{id}")
     public String updateCompany(@ModelAttribute("company") Company company, Model model) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         Company found = companyRepo.findByName(company.getName());
 
         if ((found != null) && (found.getId() != company.getId())) {
@@ -61,7 +78,20 @@ public class CompanyController { // (!) still available by usual User
 
     @GetMapping("/delete/{id}")
     public String deleteCompany(@PathVariable("id") Long id) {
+        // @PreAuthorize("hasRole('ROLE_ADMIN')") didn't work
+        if (!isCurrentUserAdmin()) throw new RuntimeException("not admin");
+
         companyRepo.deleteById(id);
         return "redirect:/companies";
+    }
+
+    private boolean isCurrentUserAdmin() {
+        if (((org.springframework.security.core.userdetails.User) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal())
+                .getAuthorities()
+                .contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) return true;
+        return false;
     }
 }
